Fix flagging button
